docs: added KR operators up to Kirin R Yato
In KR, There was a problem that certain separation operators were not
recognized by OCR, so we modified it to work by not recognizing existing
operators again. like the last episode of "knight (나이트)".
The following operators are not recognizing anymore. Yato, Noir Corne.
Because alternative operators are more likely to be used in the R.I.I.C

---------

Co-authored-by: David <56174894+Constrat@users.noreply.github.com>